+++ changelog.md
@@ -13,6 +13,7 @@
 ## Improvements
 
 * For GDS admin users, `gds.graph.get` is now able to resolve graph names into `Graph` objects of other users graph projections.
+* Add support for `gds.alpha.triangles`.
 
 
 ## Other changes